resolve minor issues and adapt tests to new signature
